+++ src/execution_plans/codec.rs
@@ -1,6 +1,6 @@
 use super::PartitionIsolatorExec;
 use crate::common::ComposedPhysicalExtensionCodec;
-use crate::plan::arrow_flight_read::ArrowFlightReadExec;
+use crate::execution_plans::arrow_flight_read::ArrowFlightReadExec;
 use crate::user_codec_ext::get_distributed_user_codec;
 use datafusion::arrow::datatypes::Schema;
 use datafusion::execution::FunctionRegistry;
