+++ src/execution_plans/mod.rs
@@ -1,7 +1,7 @@
 mod arrow_flight_read;
 mod codec;
-mod isolator;
+mod partition_isolator;
 
 pub use arrow_flight_read::ArrowFlightReadExec;
 pub use codec::DistributedCodec;
-pub use isolator::{PartitionGroup, PartitionIsolatorExec};
+pub use partition_isolator::{PartitionGroup, PartitionIsolatorExec};
